+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add `ShowIfWrapperComponent` for wrappers that are conditional on multiple child components
+
 ## 0.2.4 (2025-01-03)
 
 - Add inheritance strategies to style variants ([@omarluq][])