+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 - Add `#wrapped_in` helper to support wrappers that are conditional on multiple child components
+- Add `#placeholder` method to `ViewComponentContrib::WrapperComponent` that only renders a block if no conditional child components render
 
 ## 0.2.4 (2025-01-03)