+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## master
 
-- Fix compatibility with view_component v4. 
+## 0.2.5 (2025-08-11)
+
+- Fix compatibility with view_component v4.
 
 ## 0.2.4 (2025-01-03)